added some docs
